recommendation_service/load_catalog/requirements.txt
Werkzeug==2.3.7
Flask==2.2.2
gunicorn==20.1.0
flask-cors
pandas
pandas-gbq
db-dtypes
gcsfs
faker
google-cloud-alloydb-connector[pg8000]
sqlalchemy
google-cloud-storage
functions-framework
gradio
google-cloud-aiplatform
scikit-learn
jsonpickle